+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to the "Dev-Store" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.3.0] - 2025
+### Added
+- Environment variable support in commands using {env:VARIABLE} syntax
+- Automatic .env file detection and variable replacement
+- Warning notifications for missing environment variables
+
+## [1.2.0] - 2025
+### Added
+- File placeholder support in commands
+- Placeholders for {filename} and {filepath}
+- Automatic replacement of placeholders with active file information
+
 ## [1.1.1] - 2025
 ### Added
 - Initial release of Dev Store